+++ packages/dnr-rulesets/tasks/build.ts
@@ -4,7 +4,9 @@ import path from 'path';
 
 import {
     BASE_DIR,
+    BrowserFilters,
     DEST_RULESETS_DIR,
+    FILTERS_BROWSER_STUB,
     FILTERS_DIR,
     LOCAL_METADATA_FILE_NAME,
     RESOURCES_DIR,
@@ -59,26 +61,33 @@ const removeFiltersMetadata = async (dir: string): Promise<void> => {
 
 /**
  * Compiles rules to declarative json
- * Actually for each rule set entry in manifest's declarative_net_request:
+ * Actually for each rule set entry in manifest's declarative_net_request.
  *
+ * ```json
  * "declarative_net_request": {
  *   "rule_resources": [{
  *     "id": "ruleset_1",
  *     "enabled": true,
  *     "path": "filters/declarative/rules.json"
  *   }]
- * }.
+ * }
+ * ```
  *
  * We should find corresponding text file in resources,
  * and then convert and save json to path specified in the manifest.
+ *
+ * @param browser Browser to build filters for.
  */
-const build = async (): Promise<void> => {
-    await startDownload();
+const build = async (browser: BrowserFilters): Promise<void> => {
+    const filtersDir = FILTERS_DIR.replace(FILTERS_BROWSER_STUB, browser);
+    const destRulesetsDir = DEST_RULESETS_DIR.replace(FILTERS_BROWSER_STUB, browser);
+
+    await startDownload(filtersDir, browser);
 
     await convertFilters(
-        FILTERS_DIR,
+        filtersDir,
         RESOURCES_DIR,
-        DEST_RULESETS_DIR,
+        destRulesetsDir,
         {
             debug: true,
             prettifyJson: false,
@@ -89,13 +98,20 @@ const build = async (): Promise<void> => {
         },
     );
 
-    await removeTxtFiles(FILTERS_DIR);
+    await removeTxtFiles(filtersDir);
 
     // After single conversion, do not forget to remove filters.json file,
     // since it is packed inside metadata ruleset.
-    await removeFiltersMetadata(FILTERS_DIR);
+    await removeFiltersMetadata(filtersDir);
+};
 
+/**
+ * Builds all filters for all browsers.
+ */
+const buildAll = async (): Promise<void> => {
+    await build(BrowserFilters.ChromiumMV3);
+    await build(BrowserFilters.Opera);
     await createVersionTxt();
 };
 
-build();
+buildAll();